Move activity and view model to a new package
